Extract contribution domain into service module
