+++ services/contribution/src/main/java/org/open4goods/services/contribution/config/yml/ContributedOrganisation.java
@@ -1,0 +1,61 @@
+package org.open4goods.services.contribution.config.yml;
+
+import jakarta.validation.constraints.NotEmpty;
+
+public class ContributedOrganisation {
+
+    @NotEmpty
+    private String name;
+
+    @NotEmpty
+    private String img;
+
+    @NotEmpty
+    private String url;
+
+    @NotEmpty
+    private String description;
+
+    @NotEmpty
+    private String shortDescription;
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getImg() {
+        return img;
+    }
+
+    public void setImg(String img) {
+        this.img = img;
+    }
+
+    public String getUrl() {
+        return url;
+    }
+
+    public void setUrl(String url) {
+        this.url = url;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    public String getShortDescription() {
+        return shortDescription;
+    }
+
+    public void setShortDescription(String shortDescription) {
+        this.shortDescription = shortDescription;
+    }
+}
