+++ services/contribution/src/main/java/org/open4goods/services/contribution/repository/ContributionVoteRepository.java
@@ -1,22 +1,18 @@
-package org.open4goods.ui.repository;
+package org.open4goods.services.contribution.repository;
 
-import org.open4goods.commons.model.data.ContributionVote;
+import org.open4goods.services.contribution.model.ContributionVote;
 import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
 
 /**
  * Spring repository for contribution vote
  */
-public interface  ContributionVoteRepository extends ElasticsearchRepository<ContributionVote, String> {
+public interface ContributionVoteRepository extends ElasticsearchRepository<ContributionVote, String> {
 
-	
-	
-	  /**
+    /**
      * Counts all occurrences of ContributionVote from a given date based on the "ts" field.
-     * 
+     *
      * @param ts the timestamp from which to start counting
      * @return the number of occurrences
      */
     long countByTsGreaterThanEqual(long ts);
-    
-    
 }
