+++ ui/pom.xml
@@ -163,6 +163,12 @@
             <version>${global.version}</version>
         </dependency>
 
+        <dependency>
+            <groupId>org.open4goods</groupId>
+            <artifactId>contribution</artifactId>
+            <version>${global.version}</version>
+        </dependency>
+
 
 
         <!-- Commons  -->
