+++ ui/src/main/java/org/open4goods/ui/config/AppConfig.java
@@ -42,8 +42,8 @@ import org.open4goods.services.urlfetching.service.UrlFetchingService;
 import org.open4goods.ui.config.yml.UiConfig;
 import org.open4goods.ui.interceptors.GenericTemplateInterceptor;
 import org.open4goods.ui.interceptors.ImageResizeInterceptor;
-import org.open4goods.ui.repository.ContributionVoteRepository;
-import org.open4goods.ui.services.ContributionService;
+import org.open4goods.services.contribution.repository.ContributionVoteRepository;
+import org.open4goods.services.contribution.service.ContributionService;
 import org.open4goods.ui.services.GoogleIndexationService;
 import org.open4goods.services.gtinservice.service.GtinService;
 import org.open4goods.ui.services.SitemapGenerationService;
