+++ ui/src/main/java/org/open4goods/ui/controllers/ui/pages/PartenairesController.java
@@ -6,7 +6,7 @@ import java.util.HashMap;
 import java.util.Map;
 
 import org.open4goods.commons.config.yml.datasource.DataSourceProperties;
-import org.open4goods.commons.model.data.ContributionVote;
+import org.open4goods.services.contribution.model.ContributionVote;
 import org.open4goods.commons.services.DataSourceConfigService;
 import org.open4goods.services.serialisation.service.SerialisationService;
 import org.open4goods.ui.config.yml.UiConfig;
